feat: Phase 6c - ECO training pipeline integration + landscape results (ARI up to 0.47 mean)
Integrate ECOClusterLoss into the training pipeline as j-invariant
regularization alongside DirectCluster. Key changes:

- Add --use-eco CLI flag to train_2d.py with joint
  DirectCluster + ECO loss formulation
- Fix BF16 compatibility in eco_cluster.py metric logging
- Run 8-seed landscape scan on Tesla T4 confirming ECO
  improves low-end seeds (mean ARI 0.421 -> 0.470) but
  variance compression is limited (sigma 0.390 -> 0.369)

+++ src/losses/eco_cluster.py
@@ -361,11 +361,11 @@ class ECOClusterLoss(nn.Module):
             'cluster_balance': (cluster_sizes.min() / cluster_sizes.max().clamp(min=1)).item(),
             'row_entropy': row_ent.item(),
             'P_max_mean': P.max(dim=1)[0].mean().item(),
-            'j_current': j_cur.detach().cpu().numpy(),
-            'j_initial': self.j_initial.detach().cpu().numpy(),
+            'j_current': j_cur.detach().float().cpu().numpy(),
+            'j_initial': self.j_initial.detach().float().cpu().numpy(),
             'j_drift': (j_cur - self.j_initial).abs().mean().item(),
-            'curve_a': a.detach().cpu().numpy(),
-            'curve_b': b.detach().cpu().numpy(),
+            'curve_a': a.detach().float().cpu().numpy(),
+            'curve_b': b.detach().float().cpu().numpy(),
         }
 
         return loss, P, metrics
